Refactor Python DSL v2 to unified AST rewriter and structured IR
- Unified AST rewriter in rewriter.py to generate IR-building code.\n- Merged JIT and runtime into multistage.py supporting static and dynamic stages.\n- Replaced BR/COND_BR with structured IF, LOOP, and SWITCH operations.\n- Updated PrettyPrinter to use C-like style for structured control flow.\n- Renamed unrolled(range(n)) to static_range(n) and added StaticIf/StaticWhile.\n- Added support for Python native match statement translation to IR SWITCH.\n- Verified with 135 tests including new polymorphic dispatch example.

+++ python/tests/test_utils.py
@@ -54,9 +54,9 @@ def test_unrolled_builds_ir():
     
     @callable
     def sum_unrolled() -> int32:
-        total = 0
+        total = int32(0)
         for i in unrolled(range(4)):
-            total = total + i
+            total = total + int32(i)
         return total
     
     ir = sum_unrolled()
@@ -136,10 +136,10 @@ def test_nested_unrolled():
     
     @callable
     def nested_sum() -> int32:
-        total = 0
+        total = int32(0)
         for i in unrolled(range(3)):
             for j in unrolled(range(3)):
-                total = total + i + j
+                total = total + int32(i) + int32(j)
         return total
     
     ir = nested_sum()
